Some logical error fixed
# Bug in sort()
In the sorting algorithm, negative number was been used while the size_t (unsigned long) is used to as the type the variables which doesn't support negative number, and therefore caused endless loop.
The bug is fixed by using type long long (signed long long) instead of size_t (unsigned long).

# Bug in the Disk scheduling algorithm
The for-loop in the functions scan_alg(..), c_scan_alg(...), c_look_alg(...) has a small bug with the output associating with the input, they should be count in different index but it shared the same. New variable is introduced for counting the index of the input array.

#Bug in the displaying and array
sizeof(array) is only able to work with fixed array, size of the array (pointer) using malloc(...) cannot be count using sizeof(...). The initial size used in malloc(...) was been stored with variable in case the size is needed at some where else (In functions, parameter added for getting the size of the array passed).